Placeholder: journal hook integration
